added update_leaderboard_command, removed leaderboard.py
